requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.2.0
